requirements.txt
# Core Requirments
RPi.GPIO==0.7.0
Flask==1.1.2
gunicorn==20.0.4
python-dotenv==0.15.0

# Optional
# rtttl
pip-review

# Supporting
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
